Use `AnsiColor.CYAN` for messages not related to test results
Let's use red, green, and yellow only for messages that express test
results, and use a different color for highlighting other
informational messages.

Why `AnsiColor.CYAN`? For me, cyan is a good neutral color that is
pretty legible on both a light and dark themed terminal window. By
contrast, `BLUE` and `B_BLUE` are not very legible with a dark theme,
and `MAGENTA` looks too much like red, possibly leading people to
think that it represents a failure.